make WERROR=1, and CI builds with it; fix the sign-compare it caught
The nidx bound added in 771f0c0 compared an int with lengthof()'s size_t.
Clean under -Werror against assert-enabled and release installs of 16, 17,
18, 19 and master, extension and unit tests both.

+++ src/lion_count.c
@@ -4869,7 +4869,7 @@ lion_count_open_indexes(Snapshot snapshot, int nidx, const Oid *idxoid,
 	int			i;
 
 	/* index[] and keytype[] hold two; every caller opens one or two */
-	Assert(nidx >= 1 && nidx <= lengthof(call->index));
+	Assert(nidx >= 1 && nidx <= (int) lengthof(call->index));
 
 	call->nkeys = nidx;
 	call->heap = NULL;
